apps/crystallization-ml-service/requirements.txt
numpy==1.24.3
tensorflow==2.15.0
keras==2.15.0
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
python-dateutil==2.8.2
